Seperated the profile.js file
